refactor(parser): split parser/types.rs by responsibility (#1127)
Behavior-preserving module split of the ~2,856-line parser/types.rs into
parser/types/{mod,primitives,collections,udt,tombstones}.rs per epic #1116.

- primitives.rs: scalar parsers (int/text/uuid/timestamp/decimal/duration/inet/...)
- collections.rs: list/set/map (legacy, v5, registry-aware, schema-driven)
- udt.rs: UDT, tuple, and frozen decoding
- tombstones.rs: tombstone/deletion decoding
- mod.rs: shared model (CqlTypeId, type-id<->value mapping, empty-value
  construction), value-dispatch entry points, schema dispatcher, buffer-
  consumption helpers, serialization; re-exports the full public surface
  unchanged via pub use, so parser::types::* callers are unaffected.

Tests follow their code into per-concern #[cfg(test)] modules (44 tests
preserved, 1:1). De-duplicated the inline MAX_COLLECTION_SIZE literal into a
single const in collections.rs (no behavior change). No public API changes;
no parent mod edits required (mod types; auto-resolves to types/mod.rs).
