--- cqlite-core/src/parser/types.rs
@@ -1,2856 +1,0 @@
-//! CQL type system parsing and serialization
-//!
-//! This module handles parsing and serialization of CQL data types as they
-//! appear in Cassandra SSTable format. It maps between the binary representation
-//! and the CQLite Value types.
-
-use super::vint::{encode_vint, parse_vint, parse_vint_length};
-use crate::{
-    error::{Error, Result},
-    schema::{CqlType, UdtRegistry},
-    types::{RowKey, TombstoneInfo, TombstoneType, UdtField, UdtTypeDef, UdtValue, Value},
-};
-use nom::{
-    bytes::complete::take,
-    combinator::{map, map_res},
-    number::complete::{be_f32, be_f64, be_i32, be_i64, be_u16, be_u32, be_u8},
-    IResult,
-};
-
-/// CQL type identifiers as they appear in the binary format
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u8)]
-pub enum CqlTypeId {
-    Custom = 0x00,
-    Ascii = 0x01,
-    BigInt = 0x02,
-    Blob = 0x03,
-    Boolean = 0x04,
-    Counter = 0x05,
-    Decimal = 0x06,
-    Double = 0x07,
-    Float = 0x08,
-    Int = 0x09,
-    Timestamp = 0x0B,
-    Uuid = 0x0C,
-    Varchar = 0x0D,
-    Varint = 0x0E,
-    Timeuuid = 0x0F,
-    Inet = 0x10,
-    Date = 0x11,
-    Time = 0x12,
-    Smallint = 0x13,
-    Tinyint = 0x14,
-    Duration = 0x15,
-    List = 0x20,
-    Map = 0x21,
-    Set = 0x22,
-    Udt = 0x30,
-    Tuple = 0x31,
-    // Custom CQLite extension for tombstones
-    Tombstone = 0xFF,
-}
-
-impl TryFrom<u8> for CqlTypeId {
-    type Error = Error;
-
-    fn try_from(value: u8) -> Result<Self> {
-        match value {
-            0x00 => Ok(CqlTypeId::Custom),
-            0x01 => Ok(CqlTypeId::Ascii),
-            0x02 => Ok(CqlTypeId::BigInt),
-            0x03 => Ok(CqlTypeId::Blob),
-            0x04 => Ok(CqlTypeId::Boolean),
-            0x05 => Ok(CqlTypeId::Counter),
-            0x06 => Ok(CqlTypeId::Decimal),
-            0x07 => Ok(CqlTypeId::Double),
-            0x08 => Ok(CqlTypeId::Float),
-            0x09 => Ok(CqlTypeId::Int),
-            0x0B => Ok(CqlTypeId::Timestamp),
-            0x0C => Ok(CqlTypeId::Uuid),
-            0x0D => Ok(CqlTypeId::Varchar),
-            0x0E => Ok(CqlTypeId::Varint),
-            0x0F => Ok(CqlTypeId::Timeuuid),
-            0x10 => Ok(CqlTypeId::Inet),
-            0x11 => Ok(CqlTypeId::Date),
-            0x12 => Ok(CqlTypeId::Time),
-            0x13 => Ok(CqlTypeId::Smallint),
-            0x14 => Ok(CqlTypeId::Tinyint),
-            0x15 => Ok(CqlTypeId::Duration),
-            0x20 => Ok(CqlTypeId::List),
-            0x21 => Ok(CqlTypeId::Map),
-            0x22 => Ok(CqlTypeId::Set),
-            0x30 => Ok(CqlTypeId::Udt),
-            0x31 => Ok(CqlTypeId::Tuple),
-            0xFF => Ok(CqlTypeId::Tombstone),
-            _ => Err(Error::corruption(format!(
-                "Unknown CQL type ID: 0x{:02X}",
-                value
-            ))),
-        }
-    }
-}
-
-/// Parse a CQL type identifier
-pub fn parse_cql_type_id(input: &[u8]) -> IResult<&[u8], CqlTypeId> {
-    map_res(be_u8, CqlTypeId::try_from)(input)
-}
-
-/// Parse a CQL value based on its type with enhanced collection support
-pub fn parse_cql_value(input: &[u8], type_id: CqlTypeId) -> IResult<&[u8], Value> {
-    match type_id {
-        CqlTypeId::Boolean => parse_boolean(input),
-        CqlTypeId::Tinyint => parse_tinyint(input),
-        CqlTypeId::Smallint => parse_smallint(input),
-        CqlTypeId::Int => parse_int(input),
-        CqlTypeId::BigInt => parse_bigint(input),
-        CqlTypeId::Counter => parse_counter(input),
-        CqlTypeId::Float => parse_float(input),
-        CqlTypeId::Double => parse_double(input),
-        CqlTypeId::Ascii | CqlTypeId::Varchar => {
-            // In the SSTable cell format the caller has already extracted exactly
-            // the bytes belonging to this cell (length-framing is done at the cell
-            // level before parse_cql_value is invoked).  The entire `input` slice
-            // IS the text value — no additional length prefix to strip.
-            //
-            // For Ascii, Cassandra's AsciiSerializer requires every byte to be in
-            // the 0x00-0x7F range.  For Varchar/Text any valid UTF-8 is accepted.
-            // We accept both under the same validation path here; a stricter
-            // Ascii-only byte check can be added when needed.
-            #[cfg(feature = "legacy-heuristics")]
-            {
-                // Legacy path: try 4-byte big-endian length prefix, then null-terminated,
-                // then raw UTF-8 (kept for pre-5.0 compatibility only).
-                if input.len() >= 4 {
-                    let length =
-                        u32::from_be_bytes([input[0], input[1], input[2], input[3]]) as usize;
-                    if input.len() >= 4 + length {
-                        let text_bytes = &input[4..4 + length];
-                        if let Ok(text) = String::from_utf8(text_bytes.to_vec()) {
-                            return Ok((&input[4 + length..], Value::Text(text)));
-                        }
-                    }
-                }
-                if let Some(null_pos) = input.iter().position(|&b| b == 0) {
-                    if let Ok(text) = String::from_utf8(input[..null_pos].to_vec()) {
-                        return Ok((&input[null_pos + 1..], Value::Text(text)));
-                    }
-                }
-                if let Ok(text) = String::from_utf8(input.to_vec()) {
-                    return Ok((&[], Value::Text(text)));
-                }
-                parse_text(input)
-            }
-            #[cfg(not(feature = "legacy-heuristics"))]
-            {
-                // Deterministic path: treat the entire input as the UTF-8 text value.
-                let text = String::from_utf8(input.to_vec()).map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?;
-                Ok((&[], Value::Text(text)))
-            }
-        }
-        CqlTypeId::Blob => {
-            // For test compatibility, if input is exactly the expected size without length prefix, return it as-is
-            if input.len() == 16
-                && input
-                    == [
-                        0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x0A, 0x0B,
-                        0x0C, 0x0D, 0x0E, 0x0F,
-                    ]
-            {
-                return Ok((&[], Value::Blob(input.to_vec())));
-            }
-            // Try 4-byte big-endian length prefix first (for test compatibility)
-            if input.len() >= 4 {
-                let length = u32::from_be_bytes([input[0], input[1], input[2], input[3]]) as usize;
-                if input.len() >= 4 + length {
-                    let blob_bytes = &input[4..4 + length];
-                    return Ok((&input[4 + length..], Value::Blob(blob_bytes.to_vec())));
-                }
-            }
-            // Fallback to VInt parsing
-            parse_blob(input)
-        }
-        CqlTypeId::Uuid | CqlTypeId::Timeuuid => parse_uuid(input),
-        CqlTypeId::Timestamp => parse_timestamp(input),
-        CqlTypeId::Date => parse_date(input),
-        CqlTypeId::Time => parse_time(input),
-        CqlTypeId::Varint => parse_varint(input),
-        CqlTypeId::Decimal => parse_decimal(input),
-        CqlTypeId::Duration => parse_duration(input),
-        CqlTypeId::Inet => parse_inet(input),
-        CqlTypeId::List => parse_list_enhanced(input),
-        CqlTypeId::Set => parse_set_enhanced(input),
-        CqlTypeId::Map => parse_map_enhanced(input),
-        CqlTypeId::Udt => parse_udt_enhanced(input),
-        CqlTypeId::Tuple => parse_tuple(input),
-        CqlTypeId::Tombstone => parse_tombstone(input),
-        CqlTypeId::Custom => {
-            // Custom types require additional metadata, return as blob for now
-            parse_blob(input)
-        }
-    }
-}
-
-/// Parse a CQL value from raw data (without length prefix)
-/// This is used for collection elements where length is already parsed
-pub fn parse_cql_value_raw(input: &[u8], type_id: CqlTypeId) -> IResult<&[u8], Value> {
-    match type_id {
-        CqlTypeId::Boolean => parse_boolean(input),
-        CqlTypeId::Tinyint => parse_tinyint(input),
-        CqlTypeId::Smallint => parse_smallint(input),
-        CqlTypeId::Int => parse_int(input),
-        CqlTypeId::BigInt => parse_bigint(input),
-        CqlTypeId::Counter => parse_counter(input),
-        CqlTypeId::Float => parse_float(input),
-        CqlTypeId::Double => parse_double(input),
-        CqlTypeId::Ascii | CqlTypeId::Varchar => {
-            // For map/collection contexts, the input is already length-prefixed at the collection level
-            // So we can treat all input as the text content directly
-            let text = String::from_utf8(input.to_vec()).map_err(|_| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?;
-            Ok((&[], Value::Text(text)))
-        }
-        CqlTypeId::Blob => {
-            // For blob, use all input as blob data
-            Ok((&[], Value::Blob(input.to_vec())))
-        }
-        CqlTypeId::Uuid | CqlTypeId::Timeuuid => parse_uuid(input),
-        CqlTypeId::Timestamp => parse_timestamp(input),
-        CqlTypeId::Date => parse_date(input),
-        CqlTypeId::Time => parse_time(input),
-        CqlTypeId::Varint => parse_varint(input),
-        CqlTypeId::Decimal => parse_decimal(input),
-        CqlTypeId::Duration => parse_duration(input),
-        CqlTypeId::Inet => parse_inet(input),
-        // Collections and complex types should not be called with raw parsing
-        CqlTypeId::List | CqlTypeId::Set | CqlTypeId::Map | CqlTypeId::Udt | CqlTypeId::Tuple => {
-            parse_cql_value(input, type_id) // Fallback to normal parsing
-        }
-        CqlTypeId::Tombstone => parse_tombstone(input),
-        CqlTypeId::Custom => {
-            // Custom types require additional metadata, return as blob for now
-            Ok((&[], Value::Blob(input.to_vec())))
-        }
-    }
-}
-
-/// Parse a boolean value (1 byte: 0x00 = false, 0x01 = true)
-pub fn parse_boolean(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_u8, |b| Value::Boolean(b != 0))(input)
-}
-
-/// Parse a tinyint (signed 8-bit integer)
-pub fn parse_tinyint(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_u8, |b| Value::Integer(b as i8 as i32))(input)
-}
-
-/// Parse a smallint (signed 16-bit integer)
-pub fn parse_smallint(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_u16, |b| Value::Integer(b as i16 as i32))(input)
-}
-
-/// Parse an int (signed 32-bit integer)
-pub fn parse_int(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_i32, Value::Integer)(input)
-}
-
-/// Parse a bigint (signed 64-bit integer)
-pub fn parse_bigint(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_i64, Value::BigInt)(input)
-}
-
-/// Parse a counter (signed 64-bit integer with counter semantics)
-pub fn parse_counter(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_i64, Value::Counter)(input)
-}
-
-/// Parse a float (32-bit floating point)
-pub fn parse_float(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_f32, Value::Float32)(input)
-}
-
-/// Parse a double (64-bit floating point)
-pub fn parse_double(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_f64, Value::Float)(input)
-}
-
-/// Parse text (length-prefixed UTF-8 string)
-pub fn parse_text(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, length) = parse_vint_length(input)?;
-    let (input, bytes) = take(length)(input)?;
-    let text = String::from_utf8(bytes.to_vec()).map_err(|_| {
-        nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-    })?;
-    Ok((input, Value::Text(text)))
-}
-
-/// Parse blob (length-prefixed binary data)
-pub fn parse_blob(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, length) = parse_vint_length(input)?;
-    let (input, bytes) = take(length)(input)?;
-    Ok((input, Value::Blob(bytes.to_vec())))
-}
-
-/// Parse UUID (16 bytes)
-pub fn parse_uuid(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, bytes) = take(16usize)(input)?;
-    let mut uuid = [0u8; 16];
-    uuid.copy_from_slice(bytes);
-    Ok((input, Value::Uuid(uuid)))
-}
-
-/// Parse timestamp from binary format (milliseconds since Unix epoch)
-pub fn parse_timestamp(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_i64, Value::Timestamp)(input)
-}
-
-/// Parse date (32-bit days since epoch with Integer.MIN_VALUE offset)
-///
-/// Cassandra encodes DATE as an unsigned 32-bit integer shifted by Integer.MIN_VALUE
-/// (2^31 = 2,147,483,648) for byte-order comparability. To decode, we add i32::MIN back.
-pub fn parse_date(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_u32, |stored| {
-        // Cassandra DATE: decode by adding i32::MIN back
-        let days_since_epoch = stored.wrapping_add(i32::MIN as u32) as i32;
-        Value::Date(days_since_epoch)
-    })(input)
-}
-
-/// Parse time (64-bit nanoseconds since midnight)
-pub fn parse_time(input: &[u8]) -> IResult<&[u8], Value> {
-    map(be_i64, Value::Time)(input)
-}
-
-/// Parse varint (variable-length integer)
-pub fn parse_varint(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, length) = parse_vint_length(input)?;
-    let (input, bytes) = take(length)(input)?;
-    Ok((input, Value::Varint(bytes.to_vec())))
-}
-
-/// Parse decimal (scale + unscaled BigInteger bytes)
-///
-/// Cassandra SSTable format (DecimalType.java:275-278):
-///   - 4 bytes: scale as big-endian signed int32
-///   - remaining bytes: unscaled value as raw BigInteger two's-complement big-endian bytes
-///
-/// The unscaled bytes are NOT VInt-encoded; they are the raw Java BigInteger
-/// byte array written by `ByteBuffer.putInt(scale)` then `unscaledValue.toByteArray()`.
-pub fn parse_decimal(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, scale) = be_i32(input)?;
-    // All remaining bytes in this cell's data are the BigInteger unscaled bytes.
-    // In the SSTable cell framing, each cell value is already length-delimited by
-    // the surrounding cell header, so we consume all of `input` here.
-    let unscaled = input.to_vec();
-    Ok((
-        &input[input.len()..], // empty remaining slice
-        Value::Decimal { scale, unscaled },
-    ))
-}
-
-/// Parse duration (months, days, nanoseconds)
-///
-/// Cassandra SSTable format (DurationType.java / DurationSerializer):
-///   - months:  signed VInt (zigzag-encoded)
-///   - days:    signed VInt (zigzag-encoded)
-///   - nanos:   signed VInt (zigzag-encoded)
-///
-/// All three components are returned as-is in `Value::Duration` to preserve
-/// calendar semantics (months ≠ 30 days; days ≠ 24 hours in all time zones).
-pub fn parse_duration(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, months) = parse_vint(input)?;
-    let (input, days) = parse_vint(input)?;
-    let (input, nanos) = parse_vint(input)?;
-
-    Ok((
-        input,
-        Value::Duration {
-            months: months as i32,
-            days: days as i32,
-            nanos,
-        },
-    ))
-}
-
-/// Parse inet address (4 or 16 bytes)
-pub fn parse_inet(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, length) = parse_vint_length(input)?;
-    let (input, bytes) = take(length)(input)?;
-    Ok((input, Value::Inet(bytes.to_vec())))
-}
-
-/// Parse list using enhanced Cassandra 5+ parser with fallback to legacy format
-pub fn parse_list_enhanced(input: &[u8]) -> IResult<&[u8], Value> {
-    // Try Cassandra 5+ format first, fall back to legacy on failure
-    match parse_list_v5_format(input) {
-        Ok(result) => Ok(result),
-        Err(_) => parse_list(input), // Fallback to legacy format
-    }
-}
-
-/// Legacy list parser for backward compatibility
-pub fn parse_list(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion attacks
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::List(Vec::new())));
-    }
-
-    let (input, element_type) = parse_cql_type_id(input)?;
-
-    let mut elements = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for _ in 0..count {
-        // Parse element length prefix using VInt (which can be negative for null)
-        let (new_remaining, element_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let element = if element_length == -1 {
-            Value::Null // Null element
-        } else if element_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, element_data) = take(element_length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            // Handle nested collections correctly
-            match element_type {
-                CqlTypeId::List => parse_list(element_data)?.1,
-                CqlTypeId::Set => parse_set(element_data)?.1,
-                CqlTypeId::Map => parse_map(element_data)?.1,
-                _ => parse_cql_value_raw(element_data, element_type)?.1,
-            }
-        };
-
-        elements.push(element);
-    }
-
-    Ok((remaining, Value::List(elements)))
-}
-
-/// Parse set using enhanced Cassandra 5+ parser with fallback to legacy format
-pub fn parse_set_enhanced(input: &[u8]) -> IResult<&[u8], Value> {
-    // Try Cassandra 5+ format first, fall back to legacy on failure
-    match parse_set_v5_format(input) {
-        Ok(result) => Ok(result),
-        Err(_) => parse_set(input), // Fallback to legacy format
-    }
-}
-
-/// Legacy set parser for backward compatibility
-pub fn parse_set(input: &[u8]) -> IResult<&[u8], Value> {
-    let (remaining, list_value) = parse_list(input)?;
-
-    if let Value::List(elements) = list_value {
-        // Convert to Set - in Cassandra, sets maintain insertion order but enforce uniqueness
-        // We preserve the order as read from the SSTable for compatibility
-        Ok((remaining, Value::Set(elements)))
-    } else {
-        Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::Verify,
-        )))
-    }
-}
-
-/// Parse map using enhanced Cassandra 5+ parser with fallback to legacy format
-pub fn parse_map_enhanced(input: &[u8]) -> IResult<&[u8], Value> {
-    // Try Cassandra 5+ format first, fall back to legacy on failure
-    match parse_map_v5_format(input) {
-        Ok(result) => Ok(result),
-        Err(_) => parse_map(input), // Fallback to legacy format
-    }
-}
-
-/// Legacy map parser for backward compatibility
-pub fn parse_map(input: &[u8]) -> IResult<&[u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::Map(Vec::new())));
-    }
-
-    let (input, key_type) = parse_cql_type_id(input)?;
-    let (input, value_type) = parse_cql_type_id(input)?;
-
-    let mut map = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for _ in 0..count {
-        // Parse key length prefix
-        let (new_remaining, key_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let key = if key_length == -1 {
-            Value::Null // Null key (unusual but possible)
-        } else if key_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, key_data) = take(key_length as usize)(remaining)?;
-            remaining = new_remaining;
-            parse_cql_value_raw(key_data, key_type)?.1
-        };
-
-        // Parse value length prefix
-        let (new_remaining, value_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let value = if value_length == -1 {
-            Value::Null // Null value
-        } else if value_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, value_data) = take(value_length as usize)(remaining)?;
-            remaining = new_remaining;
-            parse_cql_value_raw(value_data, value_type)?.1
-        };
-
-        map.push((key, value));
-    }
-
-    Ok((remaining, Value::Map(map)))
-}
-
-/// Parse UDT using enhanced parser with schema registry support
-pub fn parse_udt_enhanced(input: &[u8]) -> IResult<&[u8], Value> {
-    parse_udt_enhanced_with_registry(input, &UdtRegistry::with_cassandra5_defaults())
-}
-
-/// Parse UDT with enhanced registry support
-pub fn parse_udt_enhanced_with_registry<'a>(
-    input: &'a [u8],
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    // First, always try embedded schema parsing (most common in SSTable format)
-    match parse_udt(input) {
-        Ok((remaining, udt_value)) => {
-            // If we parsed successfully and have registry info, enhance with keyspace info
-            if let Value::Udt(ref udt) = udt_value {
-                if let Some(udt_def) = try_find_udt_in_any_keyspace(registry, &udt.type_name) {
-                    let mut enhanced_udt = udt.clone();
-                    enhanced_udt.keyspace = udt_def.keyspace.clone();
-                    return Ok((remaining, Value::Udt(enhanced_udt)));
-                }
-            }
-            Ok((remaining, udt_value))
-        }
-        Err(embedded_error) => {
-            // Embedded parsing failed, try to extract type name and use registry-based parsing
-            if let Ok((after_type_name_len, type_name_length)) = parse_vint_length(input) {
-                if let Ok((after_type_name, type_name_bytes)) =
-                    take::<_, _, nom::error::Error<&[u8]>>(type_name_length)(after_type_name_len)
-                {
-                    if let Ok(type_name) = String::from_utf8(type_name_bytes.to_vec()) {
-                        if let Some(udt_def) = try_find_udt_in_any_keyspace(registry, &type_name) {
-                            // Skip embedded schema and parse field values with registry definition
-                            if let Ok((after_schema, _)) = skip_embedded_udt_schema(after_type_name)
-                            {
-                                return parse_udt_with_schema_and_registry(
-                                    after_schema,
-                                    udt_def,
-                                    registry,
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-
-            // All advanced parsing failed, return original error
-            Err(embedded_error)
-        }
-    }
-}
-
-/// Parse UDT value with embedded schema information (for SSTable format)
-pub fn parse_udt(input: &[u8]) -> IResult<&[u8], Value> {
-    // Parse UDT type name length and name
-    let (input, type_name_length) = parse_vint_length(input)?;
-    let (input, type_name_bytes) = take(type_name_length)(input)?;
-    let type_name = String::from_utf8(type_name_bytes.to_vec()).map_err(|_| {
-        nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-    })?;
-
-    // Parse field count
-    let (input, field_count) = parse_vint_length(input)?;
-
-    // Parse field definitions (schema metadata)
-    let mut field_defs = Vec::with_capacity(field_count);
-    let mut remaining = input;
-
-    for _ in 0..field_count {
-        // Parse field name
-        let (new_remaining, field_name_length) = parse_vint_length(remaining)?;
-        let (new_remaining, field_name_bytes) = take(field_name_length)(new_remaining)?;
-        let field_name = String::from_utf8(field_name_bytes.to_vec()).map_err(|_| {
-            nom::Err::Error(nom::error::Error::new(
-                new_remaining,
-                nom::error::ErrorKind::Verify,
-            ))
-        })?;
-
-        // Parse field type ID
-        let (new_remaining, field_type_id) = parse_cql_type_id(new_remaining)?;
-
-        field_defs.push((field_name, field_type_id));
-        remaining = new_remaining;
-    }
-
-    // Parse field values
-    let mut fields = Vec::with_capacity(field_count);
-    for (field_name, field_type_id) in field_defs {
-        // Parse field length
-        let (new_remaining, length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let field_value = if length == -1 {
-            // Null field
-            None
-        } else if length == 0 {
-            // Empty field
-            Some(create_empty_value(field_type_id).map_err(|_e| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?)
-        } else {
-            // Field with data
-            let (new_remaining, field_data) = take(length as usize)(remaining)?;
-            remaining = new_remaining;
-            Some(parse_cql_value_raw(field_data, field_type_id)?.1)
-        };
-
-        fields.push(UdtField {
-            name: field_name,
-            value: field_value,
-        });
-    }
-
-    let udt = UdtValue {
-        type_name,
-        keyspace: "unknown".to_string(), // Will be resolved from schema context
-        fields,
-    };
-
-    Ok((remaining, Value::Udt(udt)))
-}
-
-/// Parse UDT value with schema context (preferred method for production)
-pub fn parse_udt_with_schema<'a>(
-    input: &'a [u8],
-    udt_def: &UdtTypeDef,
-) -> IResult<&'a [u8], Value> {
-    let mut fields = Vec::with_capacity(udt_def.fields.len());
-    let mut remaining = input;
-
-    // Parse each field according to the UDT schema definition
-    for field_def in &udt_def.fields {
-        // Parse field length
-        let (new_remaining, length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let field_value = if length == -1 {
-            // Null field
-            None
-        } else if length == 0 {
-            // Empty field - create appropriate empty value
-            Some(
-                create_empty_value_for_cql_type(&field_def.field_type).map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?,
-            )
-        } else {
-            // Field with data
-            let (new_remaining, field_data) = take(length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            // Parse field data according to its CQL type
-            Some(
-                parse_cql_value_for_type(field_data, &field_def.field_type).map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?,
-            )
-        };
-
-        fields.push(UdtField {
-            name: field_def.name.clone(),
-            value: field_value,
-        });
-    }
-
-    let udt = UdtValue {
-        type_name: udt_def.name.clone(),
-        keyspace: udt_def.keyspace.clone(),
-        fields,
-    };
-
-    Ok((remaining, Value::Udt(udt)))
-}
-
-/// Parse UDT value by looking up schema from registry with enhanced dependency resolution
-pub fn parse_udt_with_registry<'a>(
-    input: &'a [u8],
-    type_name: &str,
-    keyspace: &str,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    // First, always try embedded schema parsing (which is the most common format in SSTable data)
-    match parse_udt(input) {
-        Ok((remaining, udt_value)) => {
-            // Successful embedded parsing - check if the type name matches
-            if let Value::Udt(ref udt) = udt_value {
-                if udt.type_name == type_name {
-                    // If we have registry info, update the keyspace if needed
-                    if registry.contains_udt(keyspace, type_name) {
-                        let mut updated_udt = udt.clone();
-                        updated_udt.keyspace = keyspace.to_string();
-                        return Ok((remaining, Value::Udt(updated_udt)));
-                    }
-                }
-            }
-            Ok((remaining, udt_value))
-        }
-        Err(_) => {
-            // Embedded parsing failed, try registry-based parsing (raw field values)
-            match registry.resolve_udt_with_dependencies(keyspace, type_name) {
-                Ok(udt_def) => parse_udt_with_schema_and_registry(input, udt_def, registry),
-                Err(_) => {
-                    // Fallback: try other keyspaces (for compatibility)
-                    if let Some(udt_def) = try_find_udt_in_any_keyspace(registry, type_name) {
-                        parse_udt_with_schema_and_registry(input, udt_def, registry)
-                    } else {
-                        // Unable to parse - return the original embedded parsing error
-                        parse_udt(input)
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Find UDT in any available keyspace (fallback for missing keyspace info)
-fn try_find_udt_in_any_keyspace<'a>(
-    registry: &'a UdtRegistry,
-    type_name: &str,
-) -> Option<&'a UdtTypeDef> {
-    // Try common keyspaces in order
-    let common_keyspaces = ["system", "test_keyspace", "default", "cassandra"];
-
-    for keyspace in &common_keyspaces {
-        if let Some(udt_def) = registry.get_udt(keyspace, type_name) {
-            return Some(udt_def);
-        }
-    }
-
-    None
-}
-
-/// Skip over embedded UDT schema to get to the field values
-fn skip_embedded_udt_schema(input: &[u8]) -> IResult<&[u8], ()> {
-    // Parse field count
-    let (mut remaining, field_count) = parse_vint_length(input)?;
-
-    // Skip over field definitions (name + type for each field)
-    for _ in 0..field_count {
-        // Skip field name
-        let (new_remaining, field_name_length) = parse_vint_length(remaining)?;
-        let (new_remaining, _) = take(field_name_length)(new_remaining)?;
-
-        // Skip field type
-        let (new_remaining, _) = take(1usize)(new_remaining)?; // Type ID is 1 byte
-
-        remaining = new_remaining;
-    }
-
-    Ok((remaining, ()))
-}
-
-/// Create empty value for a CQL type
-fn create_empty_value_for_cql_type(cql_type: &CqlType) -> Result<Value> {
-    match cql_type {
-        CqlType::Boolean => Ok(Value::Boolean(false)),
-        CqlType::TinyInt => Ok(Value::TinyInt(0)),
-        CqlType::SmallInt => Ok(Value::SmallInt(0)),
-        CqlType::Int => Ok(Value::Integer(0)),
-        CqlType::BigInt => Ok(Value::BigInt(0)),
-        CqlType::Float => Ok(Value::Float32(0.0)),
-        CqlType::Double => Ok(Value::Float(0.0)),
-        CqlType::Text | CqlType::Ascii | CqlType::Varchar => Ok(Value::Text(String::new())),
-        CqlType::Blob => Ok(Value::Blob(Vec::new())),
-        CqlType::Uuid | CqlType::TimeUuid => Ok(Value::Uuid([0; 16])),
-        CqlType::Timestamp => Ok(Value::Timestamp(0)),
-        CqlType::Date => Ok(Value::Timestamp(0)),
-        CqlType::Time => Ok(Value::Timestamp(0)),
-        CqlType::List(_) => Ok(Value::List(Vec::new())),
-        CqlType::Set(_) => Ok(Value::Set(Vec::new())),
-        CqlType::Map(_, _) => Ok(Value::Map(Vec::new())),
-        CqlType::Tuple(_) => Ok(Value::Tuple(Vec::new())),
-        CqlType::Udt(name, _) => Ok(Value::Udt(UdtValue::new(
-            name.clone(),
-            "unknown".to_string(),
-        ))),
-        CqlType::Frozen(inner) => create_empty_value_for_cql_type(inner),
-        _ => Ok(Value::Null),
-    }
-}
-
-/// Parse CQL value for a specific CQL type (used for UDT fields)
-fn parse_cql_value_for_type(input: &[u8], cql_type: &CqlType) -> Result<Value> {
-    match cql_type {
-        CqlType::Frozen(inner) => {
-            let inner_value = parse_cql_value_for_type(input, inner)?;
-            Ok(Value::Frozen(Box::new(inner_value)))
-        }
-        _ => {
-            let type_id = cql_type_to_type_id(cql_type);
-            let (_, value) = parse_cql_value(input, type_id).map_err(|_| {
-                Error::corruption("Failed to parse CQL value for UDT field".to_string())
-            })?;
-            Ok(value)
-        }
-    }
-}
-
-/// Parse UDT with schema and registry support for nested UDTs
-pub fn parse_udt_with_schema_and_registry<'a>(
-    input: &'a [u8],
-    udt_def: &UdtTypeDef,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    let mut fields = Vec::with_capacity(udt_def.fields.len());
-    let mut remaining = input;
-
-    // Parse each field according to the UDT schema definition
-    for field_def in &udt_def.fields {
-        // Parse field length
-        let (new_remaining, length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let field_value = if length == -1 {
-            // Null field
-            None
-        } else if length == 0 {
-            // Empty field - create appropriate empty value
-            Some(
-                create_empty_value_for_cql_type(&field_def.field_type).map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?,
-            )
-        } else {
-            // Field with data
-            let (new_remaining, field_data) = take(length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            // Parse field data with registry support for nested UDTs
-            Some(
-                parse_cql_value_for_type_with_registry(
-                    field_data,
-                    &field_def.field_type,
-                    &udt_def.keyspace,
-                    registry,
-                )
-                .map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?,
-            )
-        };
-
-        fields.push(UdtField {
-            name: field_def.name.clone(),
-            value: field_value,
-        });
-    }
-
-    let udt = UdtValue {
-        type_name: udt_def.name.clone(),
-        keyspace: udt_def.keyspace.clone(),
-        fields,
-    };
-
-    Ok((remaining, Value::Udt(udt)))
-}
-
-/// Parse CQL value for a specific CQL type with registry support for nested UDTs
-fn parse_cql_value_for_type_with_registry(
-    input: &[u8],
-    cql_type: &CqlType,
-    keyspace: &str,
-    registry: &UdtRegistry,
-) -> Result<Value> {
-    match cql_type {
-        CqlType::Udt(udt_name, _) => {
-            // Parse nested UDT using registry
-            let (_, value) =
-                parse_udt_with_registry(input, udt_name, keyspace, registry).map_err(|_| {
-                    Error::corruption(format!("Failed to parse nested UDT '{}'", udt_name))
-                })?;
-            Ok(value)
-        }
-        CqlType::List(element_type) => {
-            // Parse list with potential UDT elements
-            let (_, value) = parse_list_with_element_type(input, element_type, keyspace, registry)
-                .map_err(|_| {
-                    Error::corruption("Failed to parse list with UDT elements".to_string())
-                })?;
-            Ok(value)
-        }
-        CqlType::Set(element_type) => {
-            // Parse set with potential UDT elements
-            let (_, value) = parse_set_with_element_type(input, element_type, keyspace, registry)
-                .map_err(|_| {
-                Error::corruption("Failed to parse set with UDT elements".to_string())
-            })?;
-            Ok(value)
-        }
-        CqlType::Map(key_type, value_type) => {
-            // Parse map with potential UDT keys/values
-            let (_, value) = parse_map_with_types(input, key_type, value_type, keyspace, registry)
-                .map_err(|_| {
-                    Error::corruption("Failed to parse map with UDT elements".to_string())
-                })?;
-            Ok(value)
-        }
-        CqlType::Frozen(inner_type) => {
-            // Parse frozen type (recursive)
-            let inner_value =
-                parse_cql_value_for_type_with_registry(input, inner_type, keyspace, registry)?;
-            Ok(Value::Frozen(Box::new(inner_value)))
-        }
-        _ => {
-            // For primitive types, use the standard parser
-            let type_id = cql_type_to_type_id(cql_type);
-            let (_, value) = parse_cql_value(input, type_id).map_err(|_| {
-                Error::corruption("Failed to parse primitive CQL value".to_string())
-            })?;
-            Ok(value)
-        }
-    }
-}
-
-/// Parse list with specific element type (including UDTs)
-fn parse_list_with_element_type<'a>(
-    input: &'a [u8],
-    element_type: &CqlType,
-    keyspace: &str,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion attacks
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::List(Vec::new())));
-    }
-
-    let mut elements = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for _ in 0..count {
-        // Parse element length
-        let (new_remaining, element_length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        if element_length > 0 {
-            let (new_remaining, element_data) = take(element_length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            let element_value = parse_cql_value_for_type_with_registry(
-                element_data,
-                element_type,
-                keyspace,
-                registry,
-            )
-            .map_err(|_| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?;
-            elements.push(element_value);
-        } else if element_length == 0 {
-            // Empty element
-            let empty_value = create_empty_value_for_cql_type(element_type).map_err(|_| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?;
-            elements.push(empty_value);
-        } else {
-            // Null element - skip for lists (null elements typically not allowed)
-            // Note: Some Cassandra versions may handle this differently
-        }
-    }
-
-    Ok((remaining, Value::List(elements)))
-}
-
-/// Parse set with specific element type (including UDTs)
-fn parse_set_with_element_type<'a>(
-    input: &'a [u8],
-    element_type: &CqlType,
-    keyspace: &str,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    let (remaining, list_value) =
-        parse_list_with_element_type(input, element_type, keyspace, registry)?;
-
-    if let Value::List(elements) = list_value {
-        // Convert to Set - in Cassandra, sets maintain insertion order but enforce uniqueness
-        Ok((remaining, Value::Set(elements)))
-    } else {
-        Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::Verify,
-        )))
-    }
-}
-
-/// Parse map with specific key and value types (including UDTs)
-fn parse_map_with_types<'a>(
-    input: &'a [u8],
-    key_type: &CqlType,
-    value_type: &CqlType,
-    keyspace: &str,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::Map(Vec::new())));
-    }
-
-    let mut map = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for _ in 0..count {
-        // Parse key
-        let (new_remaining, key_length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let key = if key_length > 0 {
-            let (new_remaining, key_data) = take(key_length as usize)(remaining)?;
-            remaining = new_remaining;
-            parse_cql_value_for_type_with_registry(key_data, key_type, keyspace, registry).map_err(
-                |_| nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify)),
-            )?
-        } else {
-            create_empty_value_for_cql_type(key_type).map_err(|_| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?
-        };
-
-        // Parse value
-        let (new_remaining, value_length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let value = if value_length > 0 {
-            let (new_remaining, value_data) = take(value_length as usize)(remaining)?;
-            remaining = new_remaining;
-            parse_cql_value_for_type_with_registry(value_data, value_type, keyspace, registry)
-                .map_err(|_| {
-                    nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-                })?
-        } else {
-            create_empty_value_for_cql_type(value_type).map_err(|_| {
-                nom::Err::Error(nom::error::Error::new(input, nom::error::ErrorKind::Verify))
-            })?
-        };
-
-        map.push((key, value));
-    }
-
-    Ok((remaining, Value::Map(map)))
-}
-
-/// Convert CqlType to CqlTypeId for parsing
-fn cql_type_to_type_id(cql_type: &CqlType) -> CqlTypeId {
-    match cql_type {
-        CqlType::Boolean => CqlTypeId::Boolean,
-        CqlType::TinyInt => CqlTypeId::Tinyint,
-        CqlType::SmallInt => CqlTypeId::Smallint,
-        CqlType::Int => CqlTypeId::Int,
-        CqlType::BigInt => CqlTypeId::BigInt,
-        CqlType::Counter => CqlTypeId::Counter,
-        CqlType::Float => CqlTypeId::Float,
-        CqlType::Double => CqlTypeId::Double,
-        CqlType::Text | CqlType::Ascii | CqlType::Varchar => CqlTypeId::Varchar,
-        CqlType::Blob => CqlTypeId::Blob,
-        CqlType::Uuid => CqlTypeId::Uuid,
-        CqlType::TimeUuid => CqlTypeId::Timeuuid,
-        CqlType::Timestamp => CqlTypeId::Timestamp,
-        CqlType::Date => CqlTypeId::Date,
-        CqlType::Time => CqlTypeId::Time,
-        CqlType::Decimal => CqlTypeId::Decimal,
-        CqlType::Duration => CqlTypeId::Duration,
-        CqlType::Inet => CqlTypeId::Inet,
-        CqlType::Varint => CqlTypeId::Varint,
-        CqlType::List(_) => CqlTypeId::List,
-        CqlType::Set(_) => CqlTypeId::Set,
-        CqlType::Map(_, _) => CqlTypeId::Map,
-        CqlType::Tuple(_) => CqlTypeId::Tuple,
-        CqlType::Udt(_, _) => CqlTypeId::Udt,
-        CqlType::Frozen(_) => CqlTypeId::Blob, // Fallback only; callers should handle Frozen explicitly
-        CqlType::Custom(_) => CqlTypeId::Blob, // Custom types as blob
-    }
-}
-
-/// Parse FROZEN<UDT> values - these are immutable UDT instances
-pub fn parse_frozen_udt<'a>(input: &'a [u8], udt_def: &UdtTypeDef) -> IResult<&'a [u8], Value> {
-    let registry = UdtRegistry::with_cassandra5_defaults();
-    let (remaining, udt_value) = parse_udt_with_schema_and_registry(input, udt_def, &registry)?;
-    Ok((remaining, Value::Frozen(Box::new(udt_value))))
-}
-
-/// Parse FROZEN<UDT> with registry support for nested dependencies
-pub fn parse_frozen_udt_with_registry<'a>(
-    input: &'a [u8],
-    udt_def: &UdtTypeDef,
-    registry: &UdtRegistry,
-) -> IResult<&'a [u8], Value> {
-    // First try to parse with embedded schema (most common case)
-    if let Ok((remaining, Value::Udt(udt_value))) = parse_udt(input) {
-        // Verify the type matches what we expect
-        if udt_value.type_name == udt_def.name {
-            let mut updated_udt = udt_value;
-            updated_udt.keyspace = udt_def.keyspace.clone();
-            return Ok((remaining, Value::Frozen(Box::new(Value::Udt(updated_udt)))));
-        }
-    }
-
-    // Fallback: try to skip embedded schema and parse with registry definition
-    if let Ok((after_type_name_len, type_name_length)) = parse_vint_length(input) {
-        if let Ok((after_type_name, _type_name_bytes)) =
-            take::<_, _, nom::error::Error<&[u8]>>(type_name_length)(after_type_name_len)
-        {
-            if let Ok((after_schema, _)) = skip_embedded_udt_schema(after_type_name) {
-                let (remaining, udt_value) =
-                    parse_udt_with_schema_and_registry(after_schema, udt_def, registry)?;
-                return Ok((remaining, Value::Frozen(Box::new(udt_value))));
-            }
-        }
-    }
-
-    // All parsing attempts failed
-    Err(nom::Err::Error(nom::error::Error::new(
-        input,
-        nom::error::ErrorKind::Verify,
-    )))
-}
-
-/// Parse tuple value according to Cassandra format specification
-pub fn parse_tuple(input: &[u8]) -> IResult<&[u8], Value> {
-    // Parse field count
-    let (input, field_count) = parse_vint_length(input)?;
-
-    // Parse field type definitions
-    let mut field_types = Vec::with_capacity(field_count);
-    let mut remaining = input;
-
-    for _ in 0..field_count {
-        let (new_remaining, field_type_id) = parse_cql_type_id(remaining)?;
-        field_types.push(field_type_id);
-        remaining = new_remaining;
-    }
-
-    // Parse field values (tuples must have exact field count, no sparse representation)
-    let mut fields = Vec::with_capacity(field_count);
-    for field_type_id in field_types {
-        // Parse field length
-        let (new_remaining, length) = be_i32(remaining)?;
-        remaining = new_remaining;
-
-        let field_value = if length == -1 {
-            Value::Null // Null field
-        } else {
-            let (new_remaining, field_data) = take(length as usize)(remaining)?;
-            remaining = new_remaining;
-            parse_cql_value_raw(field_data, field_type_id)?.1
-        };
-
-        fields.push(field_value);
-    }
-
-    Ok((remaining, Value::Tuple(fields)))
-}
-
-/// Parse tombstone information with enhanced Cassandra 5.0 compatibility
-pub fn parse_tombstone(input: &[u8]) -> IResult<&[u8], Value> {
-    // Parse deletion timestamp (microseconds since epoch)
-    let (input, deletion_time) = be_i64(input)?;
-
-    // Parse tombstone type byte
-    let (input, tombstone_type_byte) = be_u8(input)?;
-
-    let tombstone_type = match tombstone_type_byte {
-        0 => TombstoneType::RowTombstone,
-        1 => TombstoneType::CellTombstone,
-        2 => TombstoneType::RangeTombstone,
-        3 => TombstoneType::TtlExpiration,
-        _ => {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                input,
-                nom::error::ErrorKind::Verify,
-            )));
-        }
-    };
-
-    // Parse optional TTL for TTL-based tombstones
-    let (input, ttl) = if tombstone_type == TombstoneType::TtlExpiration {
-        let (input, ttl_value) = be_i64(input)?;
-        (input, Some(ttl_value))
-    } else {
-        (input, None)
-    };
-
-    // Parse optional clustering key range for range tombstones
-    let (input, range_start, range_end) = if tombstone_type == TombstoneType::RangeTombstone {
-        let (input, has_range) = be_u8(input)?;
-        if has_range != 0 {
-            let (input, start_len) = parse_vint_length(input)?;
-            let (input, start_data) = take(start_len)(input)?;
-            let (input, end_len) = parse_vint_length(input)?;
-            let (input, end_data) = take(end_len)(input)?;
-            (input, Some(start_data.to_vec()), Some(end_data.to_vec()))
-        } else {
-            (input, None, None)
-        }
-    } else {
-        (input, None, None)
-    };
-
-    let tombstone_info = TombstoneInfo {
-        deletion_time,
-        tombstone_type,
-        // The legacy CQL binary tombstone encoding carries no localDeletionTime
-        // (only deletion_time); default to 0 (#873).
-        local_deletion_time: 0,
-        ttl,
-        range_start: range_start.map(RowKey::new),
-        range_end: range_end.map(RowKey::new),
-    };
-
-    Ok((input, Value::Tombstone(tombstone_info)))
-}
-
-/// Parse List with enhanced Cassandra 5+ format supporting mixed types and proper cell handling
-pub fn parse_list_v5_format(input: &[u8]) -> IResult<&[u8], Value> {
-    // Enhanced Cassandra 5+ format with proper cell metadata handling
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion attacks
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::List(Vec::new())));
-    }
-
-    // Check if this is a mixed-type collection (Cassandra 5.0 feature)
-    let (mut remaining, format_flags) = be_u8(input)?;
-    let has_mixed_types = (format_flags & 0x01) != 0;
-
-    let mut elements = Vec::with_capacity(count);
-
-    if has_mixed_types {
-        // Each element has its own type information
-        for _ in 0..count {
-            let (new_remaining, element_type_id) = parse_cql_type_id(remaining)?;
-            let (new_remaining, element_length) = parse_vint_length(new_remaining)?;
-
-            if element_length > 0 {
-                let (new_remaining, element_data) = take(element_length)(new_remaining)?;
-                let element_value =
-                    parse_cql_value_with_cell_metadata(element_data, element_type_id).map_err(
-                        |_e| {
-                            nom::Err::Error(nom::error::Error::new(
-                                element_data,
-                                nom::error::ErrorKind::Verify,
-                            ))
-                        },
-                    )?;
-                elements.push(element_value);
-                remaining = new_remaining;
-            } else {
-                // Null element
-                elements.push(Value::Null);
-                remaining = new_remaining;
-            }
-        }
-    } else {
-        // Homogeneous collection with single type
-        let (new_remaining, element_type_id) = parse_cql_type_id(remaining)?;
-        remaining = new_remaining;
-
-        for _ in 0..count {
-            let (new_remaining, element_length) = parse_vint_length(remaining)?;
-
-            if element_length > 0 {
-                let (new_remaining, element_data) = take(element_length)(new_remaining)?;
-                let element_value =
-                    parse_cql_value_with_cell_metadata(element_data, element_type_id).map_err(
-                        |_e| {
-                            nom::Err::Error(nom::error::Error::new(
-                                element_data,
-                                nom::error::ErrorKind::Verify,
-                            ))
-                        },
-                    )?;
-                elements.push(element_value);
-                remaining = new_remaining;
-            } else {
-                // Null element
-                elements.push(Value::Null);
-                remaining = new_remaining;
-            }
-        }
-    }
-
-    Ok((remaining, Value::List(elements)))
-}
-
-/// Parse CQL value with Cassandra 5.0 cell metadata handling
-fn parse_cql_value_with_cell_metadata(input: &[u8], type_id: CqlTypeId) -> Result<Value> {
-    if input.is_empty() {
-        return Ok(Value::Null);
-    }
-
-    // Skip cell metadata if present (Cassandra 5.0 format)
-    let mut offset = 0;
-    if input.len() > 1 && (input[0] & 0x80) != 0 {
-        offset += 1; // Skip cell flags
-
-        // Skip timestamp if present (8 bytes)
-        if offset + 8 <= input.len() {
-            offset += 8;
-        }
-
-        // Skip TTL if present (4 bytes)
-        if offset < input.len() && (input[0] & 0x40) != 0 && offset + 4 <= input.len() {
-            offset += 4;
-        }
-    }
-
-    let actual_data = &input[offset..];
-    if actual_data.is_empty() {
-        return Ok(Value::Null);
-    }
-
-    // Parse the actual value
-    let (_, value) = parse_cql_value(actual_data, type_id)?;
-    Ok(value)
-}
-
-/// Parse Set with Cassandra 5+ format
-pub fn parse_set_v5_format(input: &[u8]) -> IResult<&[u8], Value> {
-    // Sets use same binary format as lists in Cassandra 5+
-    let (remaining, list_value) = parse_list_v5_format(input)?;
-
-    if let Value::List(elements) = list_value {
-        // Convert to Set - maintain insertion order for compatibility
-        Ok((remaining, Value::Set(elements)))
-    } else {
-        Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::Verify,
-        )))
-    }
-}
-
-/// Parse Map with Cassandra 5+ format (tuple-based serialization)
-pub fn parse_map_v5_format(input: &[u8]) -> IResult<&[u8], Value> {
-    // Cassandra 5+ format: [count:vint][key_type:u8][value_type:u8][pairs...]
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::Map(Vec::new())));
-    }
-
-    let (input, key_type_id) = parse_cql_type_id(input)?;
-    let (input, value_type_id) = parse_cql_type_id(input)?;
-
-    let mut pairs = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for _ in 0..count {
-        // Parse key with length prefix
-        let (new_remaining, key_length) = parse_vint_length(remaining)?;
-        let (new_remaining, key_data) = take(key_length)(new_remaining)?;
-        let (_, key) = parse_cql_value(key_data, key_type_id)?;
-
-        // Parse value with length prefix
-        let (new_remaining, value_length) = parse_vint_length(new_remaining)?;
-        let (new_remaining, value_data) = take(value_length)(new_remaining)?;
-        let (_, value) = parse_cql_value(value_data, value_type_id)?;
-
-        pairs.push((key, value));
-        remaining = new_remaining;
-    }
-
-    Ok((remaining, Value::Map(pairs)))
-}
-
-/// Create an empty value for a given type ID
-fn create_empty_value(type_id: CqlTypeId) -> Result<Value> {
-    match type_id {
-        CqlTypeId::Boolean => Ok(Value::Boolean(false)),
-        CqlTypeId::Tinyint => Ok(Value::TinyInt(0)),
-        CqlTypeId::Smallint => Ok(Value::SmallInt(0)),
-        CqlTypeId::Int => Ok(Value::Integer(0)),
-        CqlTypeId::BigInt => Ok(Value::BigInt(0)),
-        CqlTypeId::Counter => Ok(Value::Counter(0)),
-        CqlTypeId::Float => Ok(Value::Float32(0.0)),
-        CqlTypeId::Double => Ok(Value::Float(0.0)),
-        CqlTypeId::Ascii | CqlTypeId::Varchar => Ok(Value::Text(String::new())),
-        CqlTypeId::Blob => Ok(Value::Blob(Vec::new())),
-        CqlTypeId::Uuid | CqlTypeId::Timeuuid => Ok(Value::Uuid([0; 16])),
-        CqlTypeId::Timestamp => Ok(Value::Timestamp(0)),
-        CqlTypeId::List => Ok(Value::List(Vec::new())),
-        CqlTypeId::Set => Ok(Value::Set(Vec::new())),
-        CqlTypeId::Map => Ok(Value::Map(Vec::new())),
-        CqlTypeId::Tuple => Ok(Value::Tuple(Vec::new())),
-        _ => Ok(Value::Null),
-    }
-}
-
-/// Serialize a CQL value to bytes
-pub fn serialize_cql_value(value: &Value) -> Result<Vec<u8>> {
-    let mut result = Vec::new();
-
-    match value {
-        Value::Null => {
-            // Null values are represented with a special length marker
-            result.extend_from_slice(&encode_vint(-1));
-        }
-        Value::Boolean(b) => {
-            result.push(CqlTypeId::Boolean as u8);
-            result.push(if *b { 1 } else { 0 });
-        }
-        Value::Integer(i) => {
-            result.push(CqlTypeId::Int as u8);
-            result.extend_from_slice(&i.to_be_bytes());
-        }
-        Value::BigInt(i) => {
-            result.push(CqlTypeId::BigInt as u8);
-            result.extend_from_slice(&i.to_be_bytes());
-        }
-        Value::Counter(c) => {
-            result.push(CqlTypeId::Counter as u8);
-            result.extend_from_slice(&c.to_be_bytes());
-        }
-        Value::Float(f) => {
-            result.push(CqlTypeId::Double as u8);
-            result.extend_from_slice(&f.to_be_bytes());
-        }
-        Value::Text(s) => {
-            result.push(CqlTypeId::Varchar as u8);
-            result.extend_from_slice(&encode_vint(s.len() as i64));
-            result.extend_from_slice(s.as_bytes());
-        }
-        Value::Blob(b) => {
-            result.push(CqlTypeId::Blob as u8);
-            result.extend_from_slice(&encode_vint(b.len() as i64));
-            result.extend_from_slice(b);
-        }
-        Value::Timestamp(ts) => {
-            result.push(CqlTypeId::Timestamp as u8);
-            let millis = ts / 1000; // Convert microseconds to milliseconds
-            result.extend_from_slice(&millis.to_be_bytes());
-        }
-        Value::Date(days) => {
-            result.push(CqlTypeId::Date as u8);
-            result.extend_from_slice(&days.to_be_bytes());
-        }
-        Value::Time(nanos) => {
-            result.push(CqlTypeId::Time as u8);
-            result.extend_from_slice(&nanos.to_be_bytes());
-        }
-        Value::Uuid(uuid) => {
-            result.push(CqlTypeId::Uuid as u8);
-            result.extend_from_slice(uuid);
-        }
-        Value::Json(json) => {
-            // Store JSON as text
-            let json_str = json.to_string();
-            result.push(CqlTypeId::Varchar as u8);
-            result.extend_from_slice(&encode_vint(json_str.len() as i64));
-            result.extend_from_slice(json_str.as_bytes());
-        }
-        Value::List(list) => {
-            result.push(CqlTypeId::List as u8);
-            result.extend_from_slice(&encode_vint(list.len() as i64));
-
-            // For simplicity, assume all elements are the same type
-            if let Some(first) = list.first() {
-                let element_type = map_value_to_cql_type(first);
-                result.push(element_type as u8);
-
-                for element in list {
-                    if let Value::Null = element {
-                        // Null element: length = -1
-                        result.extend_from_slice(&encode_vint(-1));
-                    } else {
-                        let element_data = serialize_value_without_type_prefix(element)?;
-                        result.extend_from_slice(&encode_vint(element_data.len() as i64));
-                        result.extend_from_slice(&element_data);
-                    }
-                }
-            }
-        }
-        Value::Map(map) => {
-            result.push(CqlTypeId::Map as u8);
-            result.extend_from_slice(&encode_vint(map.len() as i64));
-
-            // For simplicity, assume all keys and values are the same type
-            if let Some((first_key, first_value)) = map.first() {
-                let key_type = map_value_to_cql_type(first_key);
-                let value_type = map_value_to_cql_type(first_value);
-                result.push(key_type as u8);
-                result.push(value_type as u8);
-
-                for (key, value) in map {
-                    // Serialize key with length prefix
-                    if let Value::Null = key {
-                        result.extend_from_slice(&encode_vint(-1));
-                    } else {
-                        let key_data = serialize_value_without_type_prefix(key)?;
-                        result.extend_from_slice(&encode_vint(key_data.len() as i64));
-                        result.extend_from_slice(&key_data);
-                    }
-
-                    // Serialize value with length prefix
-                    if let Value::Null = value {
-                        result.extend_from_slice(&encode_vint(-1));
-                    } else {
-                        let value_data = serialize_value_without_type_prefix(value)?;
-                        result.extend_from_slice(&encode_vint(value_data.len() as i64));
-                        result.extend_from_slice(&value_data);
-                    }
-                }
-            }
-        }
-        Value::TinyInt(i) => {
-            result.push(CqlTypeId::Tinyint as u8);
-            result.push(*i as u8);
-        }
-        Value::Inet(bytes) => {
-            result.push(CqlTypeId::Inet as u8);
-            result.extend_from_slice(&encode_vint(bytes.len() as i64));
-            result.extend_from_slice(bytes);
-        }
-        Value::SmallInt(i) => {
-            result.push(CqlTypeId::Smallint as u8);
-            result.extend_from_slice(&i.to_be_bytes());
-        }
-        Value::Float32(f) => {
-            result.push(CqlTypeId::Float as u8);
-            result.extend_from_slice(&f.to_be_bytes());
-        }
-        Value::Set(set) => {
-            result.push(CqlTypeId::Set as u8);
-            result.extend_from_slice(&encode_vint(set.len() as i64));
-
-            if let Some(first) = set.first() {
-                let element_type = map_value_to_cql_type(first);
-                result.push(element_type as u8);
-
-                for element in set {
-                    if let Value::Null = element {
-                        // Null element: length = -1
-                        result.extend_from_slice(&encode_vint(-1));
-                    } else {
-                        let element_data = serialize_value_without_type_prefix(element)?;
-                        result.extend_from_slice(&encode_vint(element_data.len() as i64));
-                        result.extend_from_slice(&element_data);
-                    }
-                }
-            }
-        }
-        Value::Tuple(tuple) => {
-            result.push(CqlTypeId::Tuple as u8);
-            result.extend_from_slice(&encode_vint(tuple.len() as i64));
-
-            // Serialize type information for each field
-            for element in tuple {
-                let element_type = map_value_to_cql_type(element);
-                result.push(element_type as u8);
-            }
-
-            // Serialize field values with proper length prefixes
-            for element in tuple {
-                if let Value::Null = element {
-                    // Null field: length = -1
-                    result.extend_from_slice(&(-1i32).to_be_bytes());
-                } else {
-                    let element_data = serialize_value_without_type_prefix(element)?;
-                    result.extend_from_slice(&(element_data.len() as i32).to_be_bytes());
-                    result.extend_from_slice(&element_data);
-                }
-            }
-        }
-        Value::Udt(udt) => {
-            result.push(CqlTypeId::Udt as u8);
-
-            // Serialize type name
-            result.extend_from_slice(&encode_vint(udt.type_name.len() as i64));
-            result.extend_from_slice(udt.type_name.as_bytes());
-
-            // Serialize field count
-            result.extend_from_slice(&encode_vint(udt.fields.len() as i64));
-
-            // Serialize field definitions
-            for field in &udt.fields {
-                result.extend_from_slice(&encode_vint(field.name.len() as i64));
-                result.extend_from_slice(field.name.as_bytes());
-
-                // Serialize field type (inferred from value or use blob as fallback)
-                let field_type = match &field.value {
-                    Some(value) => map_value_to_cql_type(value),
-                    None => CqlTypeId::Blob, // Null field, use generic type
-                };
-                result.push(field_type as u8);
-            }
-
-            // Serialize field values
-            for field in &udt.fields {
-                match &field.value {
-                    None => {
-                        // Null field: length = -1
-                        result.extend_from_slice(&(-1i32).to_be_bytes());
-                    }
-                    Some(value) => {
-                        let field_data = serialize_value_without_type_prefix(value)?;
-                        result.extend_from_slice(&(field_data.len() as i32).to_be_bytes());
-                        result.extend_from_slice(&field_data);
-                    }
-                }
-            }
-        }
-        Value::Frozen(boxed_value) => {
-            // For frozen values, just serialize the inner value
-            let inner_bytes = serialize_cql_value(boxed_value)?;
-            result.extend_from_slice(&inner_bytes);
-        }
-        Value::Tombstone(info) => {
-            result.push(CqlTypeId::Tombstone as u8);
-            result.extend_from_slice(&info.deletion_time.to_be_bytes());
-
-            let tombstone_type_byte = match info.tombstone_type {
-                TombstoneType::RowTombstone => 0u8,
-                TombstoneType::CellTombstone => 1u8,
-                TombstoneType::RangeTombstone => 2u8,
-                TombstoneType::TtlExpiration => 3u8,
-                // PartitionTombstone is a delta-scan concept that has no encoding
-                // in the legacy binary format.  Silently mapping it to byte 0
-                // (RowTombstone) would cause a silent lossy round-trip, violating
-                // the no-silent-corruption mandate.  Fail loudly instead.
-                TombstoneType::PartitionTombstone => {
-                    return Err(Error::invalid_operation(
-                        "PartitionTombstone cannot be serialized via the legacy CQL value \
-                         binary format; it is a delta-scan concept only",
-                    ));
-                }
-            };
-            result.push(tombstone_type_byte);
-
-            // Add TTL if present
-            if let Some(ttl) = info.ttl {
-                result.extend_from_slice(&ttl.to_be_bytes());
-            }
-
-            // Add range information for range tombstones
-            if info.tombstone_type == TombstoneType::RangeTombstone {
-                if let (Some(start), Some(end)) = (&info.range_start, &info.range_end) {
-                    result.push(1u8); // Has range marker
-                    result.extend_from_slice(&encode_vint(start.len() as i64));
-                    result.extend_from_slice(start.as_bytes());
-                    result.extend_from_slice(&encode_vint(end.len() as i64));
-                    result.extend_from_slice(end.as_bytes());
-                } else {
-                    result.push(0u8); // No range marker
-                }
-            }
-        }
-        Value::Varint(data) => {
-            result.push(CqlTypeId::Varint as u8);
-            result.extend_from_slice(&encode_vint(data.len() as i64));
-            result.extend_from_slice(data);
-        }
-        Value::Decimal { scale, unscaled } => {
-            result.push(CqlTypeId::Decimal as u8);
-            result.extend_from_slice(&scale.to_be_bytes());
-            result.extend_from_slice(&encode_vint(unscaled.len() as i64));
-            result.extend_from_slice(unscaled);
-        }
-        Value::Duration {
-            months,
-            days,
-            nanos,
-        } => {
-            result.push(CqlTypeId::Duration as u8);
-            result.extend_from_slice(&months.to_be_bytes());
-            result.extend_from_slice(&days.to_be_bytes());
-            result.extend_from_slice(&nanos.to_be_bytes());
-        }
-    }
-
-    Ok(result)
-}
-
-/// Serialize a CQL value without the type prefix byte
-fn serialize_value_without_type_prefix(value: &Value) -> Result<Vec<u8>> {
-    match value {
-        Value::Null => Ok(vec![]), // Null should be handled at higher level
-        Value::Boolean(b) => Ok(vec![if *b { 1 } else { 0 }]),
-        Value::Integer(i) => Ok(i.to_be_bytes().to_vec()),
-        Value::BigInt(i) => Ok(i.to_be_bytes().to_vec()),
-        Value::Float(f) => Ok(f.to_be_bytes().to_vec()),
-        Value::Text(s) => {
-            // For raw serialization without type prefix, just include the string bytes
-            // The length will be handled by the map format itself
-            Ok(s.as_bytes().to_vec())
-        }
-        Value::Blob(b) => {
-            // For raw serialization without type prefix, just include the blob bytes
-            // The length will be handled by the map format itself
-            Ok(b.to_vec())
-        }
-        Value::Timestamp(ts) => {
-            let millis = ts / 1000; // Convert microseconds to milliseconds
-            Ok(millis.to_be_bytes().to_vec())
-        }
-        Value::Date(days) => Ok(days.to_be_bytes().to_vec()),
-        Value::Time(nanos) => Ok(nanos.to_be_bytes().to_vec()),
-        Value::Uuid(uuid) => Ok(uuid.to_vec()),
-        Value::TinyInt(i) => Ok(vec![*i as u8]),
-        Value::Inet(bytes) => Ok(bytes.clone()),
-        Value::SmallInt(i) => Ok(i.to_be_bytes().to_vec()),
-        Value::Float32(f) => Ok(f.to_be_bytes().to_vec()),
-        // For complex types, fall back to full serialization and strip type byte
-        _ => {
-            let full_bytes = serialize_cql_value(value)?;
-            Ok(full_bytes[1..].to_vec()) // Skip the type byte
-        }
-    }
-}
-
-fn map_value_to_cql_type(value: &Value) -> CqlTypeId {
-    match value {
-        Value::Null => CqlTypeId::Blob, // Generic fallback
-        Value::Boolean(_) => CqlTypeId::Boolean,
-        Value::Integer(_) => CqlTypeId::Int,
-        Value::BigInt(_) => CqlTypeId::BigInt,
-        Value::Counter(_) => CqlTypeId::Counter,
-        Value::Float(_) => CqlTypeId::Double,
-        Value::Text(_) => CqlTypeId::Varchar,
-        Value::Blob(_) => CqlTypeId::Blob,
-        Value::Timestamp(_) => CqlTypeId::Timestamp,
-        Value::Date(_) => CqlTypeId::Date,
-        Value::Time(_) => CqlTypeId::Time,
-        Value::Uuid(_) => CqlTypeId::Uuid,
-        Value::Json(_) => CqlTypeId::Varchar,
-        Value::TinyInt(_) => CqlTypeId::Tinyint,
-        Value::Inet(_) => CqlTypeId::Inet,
-        Value::SmallInt(_) => CqlTypeId::Smallint,
-        Value::Float32(_) => CqlTypeId::Float,
-        Value::List(_) => CqlTypeId::List,
-        Value::Set(_) => CqlTypeId::Set,
-        Value::Map(_) => CqlTypeId::Map,
-        Value::Tuple(_) => CqlTypeId::Tuple,
-        Value::Udt(_) => CqlTypeId::Udt,
-        Value::Frozen(_) => CqlTypeId::Blob, // Frozen is a wrapper, use blob as fallback
-        Value::Tombstone(_) => CqlTypeId::Tombstone,
-        Value::Varint(_) => CqlTypeId::Varint,
-        Value::Decimal { .. } => CqlTypeId::Decimal,
-        Value::Duration { .. } => CqlTypeId::Duration,
-    }
-}
-
-// ============================================================================
-// Buffer Consumption Validation Helpers (Issue #61)
-// ============================================================================
-
-/// Assert that all input bytes have been consumed during parsing
-///
-/// This helper enforces complete buffer consumption to prevent silent data
-/// truncation in collection and UDT parsing. Per Issue #61 acceptance criteria,
-/// all parsers must validate full consumption.
-///
-/// # Arguments
-/// * `remaining` - The remaining bytes after parsing
-/// * `context` - Description of parsing context for error messages
-///
-/// # Returns
-/// * `Ok(())` if buffer is fully consumed
-/// * `Err` if bytes remain unconsumed
-#[inline]
-#[allow(dead_code)] // Will be used in collection parsing functions
-pub(crate) fn assert_full_buffer_consumption(remaining: &[u8], context: &str) -> Result<()> {
-    if !remaining.is_empty() {
-        return Err(Error::corruption(format!(
-            "Buffer not fully consumed in {}: {} bytes remaining",
-            context,
-            remaining.len()
-        )));
-    }
-    Ok(())
-}
-
-/// Validate that a parsed element consumed its entire allocated buffer
-///
-/// Used within collection parsing to ensure each element's buffer is fully
-/// consumed before moving to the next element.
-#[inline]
-#[allow(dead_code)] // Will be used in collection parsing functions
-pub(crate) fn validate_element_consumption<'a>(
-    _input: &'a [u8],
-    remaining: &'a [u8],
-    element_index: usize,
-    collection_type: &str,
-) -> Result<()> {
-    if !remaining.is_empty() {
-        return Err(Error::corruption(format!(
-            "{} element {} did not consume full buffer: {} bytes remaining",
-            collection_type,
-            element_index,
-            remaining.len()
-        )));
-    }
-    Ok(())
-}
-
-// ============================================================================
-// Schema-Aware Collection Parsing (Issue #61)
-// ============================================================================
-
-/// Parse list with schema-driven element decoding
-///
-/// Uses provided schema to deterministically decode list elements, ensuring
-/// no heuristic fallbacks per Issue #28. Validates full buffer consumption
-/// for each element per Issue #61.
-///
-/// # Arguments
-/// * `input` - Raw bytes to parse
-/// * `element_schema` - Schema for list elements
-///
-/// # Returns
-/// * Parsed list value with fully consumed input
-pub fn parse_list_with_schema<'a>(
-    input: &'a [u8],
-    element_schema: &CqlType,
-) -> IResult<&'a [u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::List(Vec::new())));
-    }
-
-    let mut elements = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for i in 0..count {
-        // Parse element length prefix (can be -1 for null)
-        let (new_remaining, element_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let element = if element_length == -1 {
-            Value::Null
-        } else if element_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, element_data) = take(element_length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            // Parse using schema-driven decoding (no heuristics)
-            let (element_remaining, element_value) =
-                parse_cql_value_with_schema(element_data, element_schema)?;
-
-            // Validate element consumed its full buffer (Issue #61)
-            validate_element_consumption(element_data, element_remaining, i, "List").map_err(
-                |_e| {
-                    nom::Err::Error(nom::error::Error::new(
-                        element_data,
-                        nom::error::ErrorKind::Verify,
-                    ))
-                },
-            )?;
-
-            element_value
-        };
-
-        elements.push(element);
-    }
-
-    Ok((remaining, Value::List(elements)))
-}
-
-/// Parse map with schema-driven key/value decoding
-///
-/// Uses provided schemas to deterministically decode map entries, ensuring
-/// no heuristic fallbacks per Issue #28. Validates full buffer consumption
-/// for each key and value per Issue #61.
-pub fn parse_map_with_schema<'a>(
-    input: &'a [u8],
-    key_schema: &CqlType,
-    value_schema: &CqlType,
-) -> IResult<&'a [u8], Value> {
-    let (input, count) = parse_vint_length(input)?;
-
-    // Validate count to prevent memory exhaustion
-    const MAX_COLLECTION_SIZE: usize = 1_000_000;
-    if count > MAX_COLLECTION_SIZE {
-        return Err(nom::Err::Error(nom::error::Error::new(
-            input,
-            nom::error::ErrorKind::TooLarge,
-        )));
-    }
-
-    if count == 0 {
-        return Ok((input, Value::Map(Vec::new())));
-    }
-
-    let mut pairs = Vec::with_capacity(count);
-    let mut remaining = input;
-
-    for i in 0..count {
-        // Parse key length
-        let (new_remaining, key_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let key = if key_length == -1 {
-            Value::Null
-        } else if key_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, key_data) = take(key_length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            let (key_remaining, key_value) = parse_cql_value_with_schema(key_data, key_schema)?;
-            validate_element_consumption(key_data, key_remaining, i, "Map key").map_err(|_e| {
-                nom::Err::Error(nom::error::Error::new(
-                    key_data,
-                    nom::error::ErrorKind::Verify,
-                ))
-            })?;
-
-            key_value
-        };
-
-        // Parse value length
-        let (new_remaining, value_length) = parse_vint(remaining)?;
-        remaining = new_remaining;
-
-        let value = if value_length == -1 {
-            Value::Null
-        } else if value_length < 0 {
-            return Err(nom::Err::Error(nom::error::Error::new(
-                remaining,
-                nom::error::ErrorKind::Verify,
-            )));
-        } else {
-            let (new_remaining, value_data) = take(value_length as usize)(remaining)?;
-            remaining = new_remaining;
-
-            let (value_remaining, value_value) =
-                parse_cql_value_with_schema(value_data, value_schema)?;
-            validate_element_consumption(value_data, value_remaining, i, "Map value").map_err(
-                |_e| {
-                    nom::Err::Error(nom::error::Error::new(
-                        value_data,
-                        nom::error::ErrorKind::Verify,
-                    ))
-                },
-            )?;
-
-            value_value
-        };
-
-        pairs.push((key, value));
-    }
-
-    Ok((remaining, Value::Map(pairs)))
-}
-
-/// Parse CQL value using schema (no heuristics)
-///
-/// Schema-driven parser that never falls back to heuristics. Used by
-/// schema-aware collection parsers to ensure deterministic decoding.
-fn parse_cql_value_with_schema<'a>(input: &'a [u8], schema: &CqlType) -> IResult<&'a [u8], Value> {
-    match schema {
-        CqlType::Boolean => parse_boolean(input),
-        CqlType::TinyInt => parse_tinyint(input),
-        CqlType::SmallInt => parse_smallint(input),
-        CqlType::Int => parse_int(input),
-        CqlType::BigInt => parse_bigint(input),
-        CqlType::Counter => parse_counter(input),
-        CqlType::Float => parse_float(input),
-        CqlType::Double => parse_double(input),
-        CqlType::Text | CqlType::Ascii | CqlType::Varchar => parse_text(input),
-        CqlType::Blob => parse_blob(input),
-        CqlType::Uuid | CqlType::TimeUuid => parse_uuid(input),
-        CqlType::Timestamp => parse_timestamp(input),
-        CqlType::Date => parse_date(input),
-        CqlType::Time => parse_time(input),
-        CqlType::Duration => parse_duration(input),
-        CqlType::Inet => parse_inet(input),
-        CqlType::Decimal => parse_decimal(input),
-        CqlType::Varint => parse_varint(input),
-        CqlType::List(element_type) => parse_list_with_schema(input, element_type),
-        CqlType::Set(element_type) => {
-            let (remaining, Value::List(elements)) = parse_list_with_schema(input, element_type)?
-            else {
-                return Err(nom::Err::Error(nom::error::Error::new(
-                    input,
-                    nom::error::ErrorKind::Verify,
-                )));
-            };
-            Ok((remaining, Value::Set(elements)))
-        }
-        CqlType::Map(key_type, value_type) => parse_map_with_schema(input, key_type, value_type),
-        CqlType::Tuple(_) => parse_tuple(input),
-        CqlType::Udt(_, _) => parse_udt(input),
-        CqlType::Frozen(inner) => parse_cql_value_with_schema(input, inner),
-        CqlType::Custom(_) => {
-            // Custom types require additional metadata, parse as blob
-            parse_blob(input)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_cql_type_id_conversion() {
-        assert_eq!(CqlTypeId::try_from(0x04).unwrap(), CqlTypeId::Boolean);
-        assert_eq!(CqlTypeId::try_from(0x09).unwrap(), CqlTypeId::Int);
-        assert_eq!(CqlTypeId::try_from(0xFF).unwrap(), CqlTypeId::Tombstone);
-    }
-
-    #[test]
-    fn test_boolean_parsing() {
-        let data = vec![0x01];
-        let (_, value) = parse_boolean(&data).unwrap();
-        assert_eq!(value, Value::Boolean(true));
-
-        let data = vec![0x00];
-        let (_, value) = parse_boolean(&data).unwrap();
-        assert_eq!(value, Value::Boolean(false));
-    }
-
-    #[test]
-    fn test_int_parsing() {
-        let data = vec![0x00, 0x00, 0x01, 0x00]; // 256 in big-endian
-        let (_, value) = parse_int(&data).unwrap();
-        assert_eq!(value, Value::Integer(256));
-    }
-
-    #[test]
-    fn test_text_parsing() {
-        use super::super::vint::encode_vint;
-
-        let test_str = "hello";
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(test_str.len() as i64));
-        data.extend_from_slice(test_str.as_bytes());
-
-        let (_, value) = parse_text(&data).unwrap();
-        assert_eq!(value, Value::Text("hello".to_string()));
-    }
-
-    #[test]
-    fn test_uuid_parsing() {
-        let uuid_bytes = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16];
-        let (_, value) = parse_uuid(&uuid_bytes).unwrap();
-        assert_eq!(value, Value::Uuid(uuid_bytes));
-    }
-
-    #[test]
-    fn test_value_serialization_roundtrip() {
-        let test_values = vec![
-            Value::Boolean(true),
-            Value::Integer(42),
-            Value::BigInt(1000),
-            Value::Float(std::f64::consts::PI),
-            Value::Text("test".to_string()),
-            Value::Blob(vec![1, 2, 3, 4]),
-        ];
-
-        for value in test_values {
-            let serialized = serialize_cql_value(&value).unwrap();
-            // Note: Full roundtrip testing would require implementing deserialization
-            // which depends on the type context that's not always preserved
-            assert!(!serialized.is_empty());
-        }
-    }
-
-    #[test]
-    fn test_udt_serialization() {
-        // Test UDT serialization
-        let udt = UdtValue {
-            type_name: "Person".to_string(),
-            keyspace: "test".to_string(),
-            fields: vec![
-                UdtField {
-                    name: "name".to_string(),
-                    value: Some(Value::Text("John".to_string())),
-                },
-                UdtField {
-                    name: "age".to_string(),
-                    value: Some(Value::Integer(30)),
-                },
-                UdtField {
-                    name: "email".to_string(),
-                    value: None,
-                }, // Null field
-            ],
-        };
-
-        let serialized = serialize_cql_value(&Value::Udt(udt)).unwrap();
-        assert!(!serialized.is_empty());
-
-        // Should start with UDT type ID
-        assert_eq!(serialized[0], CqlTypeId::Udt as u8);
-    }
-
-    #[test]
-    fn test_tuple_serialization() {
-        // Test tuple serialization
-        let tuple = vec![
-            Value::Text("hello".to_string()),
-            Value::Integer(42),
-            Value::Boolean(true),
-        ];
-
-        let serialized = serialize_cql_value(&Value::Tuple(tuple)).unwrap();
-        assert!(!serialized.is_empty());
-
-        // Should start with Tuple type ID
-        assert_eq!(serialized[0], CqlTypeId::Tuple as u8);
-    }
-
-    #[test]
-    fn test_tombstone_parsing() {
-        // Test row tombstone parsing
-        let row_tombstone = Value::row_tombstone(1000);
-        let serialized = serialize_cql_value(&row_tombstone).unwrap();
-
-        // Parse it back
-        let (remaining, parsed_value) = parse_tombstone(&serialized[1..]).unwrap(); // Skip type ID
-        assert!(remaining.is_empty());
-        assert_eq!(parsed_value, row_tombstone);
-
-        // Test TTL tombstone parsing
-        let ttl_tombstone = Value::ttl_tombstone(2000, 1000);
-        let serialized_ttl = serialize_cql_value(&ttl_tombstone).unwrap();
-
-        let (remaining, parsed_ttl) = parse_tombstone(&serialized_ttl[1..]).unwrap(); // Skip type ID
-        assert!(remaining.is_empty());
-        assert_eq!(parsed_ttl, ttl_tombstone);
-    }
-
-    #[test]
-    fn test_tombstone_serialization() {
-        // Test tombstone serialization
-        let tombstone = Value::cell_tombstone(5000);
-        let serialized = serialize_cql_value(&tombstone).unwrap();
-        assert!(!serialized.is_empty());
-
-        // Should start with Tombstone type ID
-        assert_eq!(serialized[0], CqlTypeId::Tombstone as u8);
-
-        // Should contain deletion time
-        let deletion_time_bytes = &serialized[1..9];
-        let deletion_time = i64::from_be_bytes([
-            deletion_time_bytes[0],
-            deletion_time_bytes[1],
-            deletion_time_bytes[2],
-            deletion_time_bytes[3],
-            deletion_time_bytes[4],
-            deletion_time_bytes[5],
-            deletion_time_bytes[6],
-            deletion_time_bytes[7],
-        ]);
-        assert_eq!(deletion_time, 5000);
-    }
-
-    #[test]
-    fn test_partition_tombstone_serialize_errors() {
-        // PartitionTombstone has no encoding in the legacy CQL binary format.
-        // Serializing one must return an explicit error, not silently alias to
-        // RowTombstone (byte 0) — that would be a silent lossy round-trip.
-        use crate::types::{TombstoneInfo, TombstoneType};
-        let partition_tombstone = Value::Tombstone(TombstoneInfo {
-            deletion_time: 9999,
-            tombstone_type: TombstoneType::PartitionTombstone,
-            local_deletion_time: 0,
-            ttl: None,
-            range_start: None,
-            range_end: None,
-        });
-        let result = serialize_cql_value(&partition_tombstone);
-        assert!(
-            result.is_err(),
-            "serializing PartitionTombstone must return Err, not Ok"
-        );
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("PartitionTombstone"),
-            "error message should mention PartitionTombstone, got: {err_msg}"
-        );
-    }
-
-    #[test]
-    fn test_parse_list_v5_homogeneous() {
-        let mut data = Vec::new();
-        data.extend(super::super::vint::encode_vint(2)); // two elements
-        data.push(0x00); // homogeneous collection
-        data.push(CqlTypeId::Int as u8);
-
-        data.extend(super::super::vint::encode_vint(4));
-        data.extend_from_slice(&1i32.to_be_bytes());
-
-        data.extend(super::super::vint::encode_vint(4));
-        data.extend_from_slice(&2i32.to_be_bytes());
-
-        let (_, value) = parse_list_v5_format(&data).expect("parse list");
-        assert_eq!(
-            value,
-            Value::List(vec![Value::Integer(1), Value::Integer(2)])
-        );
-    }
-
-    #[test]
-    fn test_parse_list_v5_mixed_types() {
-        let mut data = Vec::new();
-        data.extend(super::super::vint::encode_vint(2)); // two elements
-        data.push(0x01); // mixed-type flag
-
-        // First element: text "alpha"
-        data.push(CqlTypeId::Varchar as u8);
-        data.extend(super::super::vint::encode_vint(5));
-        data.extend_from_slice(b"alpha");
-
-        // Second element: integer 7
-        data.push(CqlTypeId::Int as u8);
-        data.extend(super::super::vint::encode_vint(4));
-        data.extend_from_slice(&7i32.to_be_bytes());
-
-        let (_, value) = parse_list_v5_format(&data).expect("parse list");
-        assert_eq!(
-            value,
-            Value::List(vec![Value::Text("alpha".to_string()), Value::Integer(7)])
-        );
-    }
-
-    // =============================================
-    // Issue #203: Additional tests for 75% coverage
-    // =============================================
-
-    #[test]
-    fn test_parse_timestamp() {
-        // Timestamp is stored as milliseconds since epoch (i64 big-endian)
-        let ts_ms: i64 = 1702900000000; // 2023-12-18 in milliseconds
-        let data = ts_ms.to_be_bytes();
-        let (remaining, value) = parse_timestamp(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Timestamp(ts_ms));
-    }
-
-    #[test]
-    fn test_parse_timestamp_negative() {
-        // Test negative timestamp (before epoch)
-        let ts_ms: i64 = -86400000; // -1 day in milliseconds
-        let data = ts_ms.to_be_bytes();
-        let (remaining, value) = parse_timestamp(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Timestamp(ts_ms));
-    }
-
-    #[test]
-    fn test_parse_varint_single_byte() {
-        // Varint with 1-byte value
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(1)); // length = 1
-        data.push(0x2A); // value 42
-        let (remaining, value) = parse_varint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Varint(vec![0x2A]));
-    }
-
-    #[test]
-    fn test_parse_varint_multi_byte() {
-        // Varint with multi-byte value (big number)
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(4)); // length = 4
-        data.extend_from_slice(&[0x01, 0x02, 0x03, 0x04]);
-        let (remaining, value) = parse_varint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Varint(vec![0x01, 0x02, 0x03, 0x04]));
-    }
-
-    #[test]
-    fn test_parse_decimal() {
-        // Cassandra decimal: 4-byte BE i32 scale + raw BigInteger bytes (NOT VInt).
-        // e.g., 1.23 = scale 2, unscaled raw bytes = [0x7B] (= 123)
-        let mut data = Vec::new();
-        data.extend_from_slice(&2i32.to_be_bytes()); // scale = 2
-        data.push(0x7B); // unscaled BigInteger bytes = 123
-        let (remaining, value) = parse_decimal(&data).unwrap();
-        assert!(remaining.is_empty());
-        match value {
-            Value::Decimal {
-                scale,
-                ref unscaled,
-            } => {
-                assert_eq!(scale, 2, "scale should be 2");
-                assert_eq!(unscaled, &[0x7B], "unscaled should be [0x7B]=123 (= 1.23)");
-            }
-            other => panic!("Expected Decimal value, got {:?}", other),
-        }
-    }
-
-    #[test]
-    fn test_parse_decimal_negative_scale() {
-        // scale = -2, unscaled = [0x05] = 5 → value is 500 (5 × 10^2)
-        let mut data = Vec::new();
-        data.extend_from_slice(&(-2i32).to_be_bytes()); // scale = -2
-        data.push(0x05); // unscaled = 5
-        let (remaining, value) = parse_decimal(&data).unwrap();
-        assert!(remaining.is_empty());
-        match value {
-            Value::Decimal {
-                scale,
-                ref unscaled,
-            } => {
-                assert_eq!(scale, -2, "scale should be -2");
-                assert_eq!(unscaled, &[0x05], "unscaled should be [0x05]=5");
-            }
-            other => panic!("Expected Decimal value, got {:?}", other),
-        }
-    }
-
-    /// S2 regression: decimal unscaled bytes > 127 must NOT be VInt-decoded.
-    /// If decoded as VInt, byte 0x80 would be misread (high bit set = multi-byte VInt).
-    #[test]
-    fn test_parse_decimal_large_unscaled_no_vint_misread() {
-        // scale = 0, unscaled = [0x01, 0x00] = 256 (big-endian BigInteger)
-        let mut data = Vec::new();
-        data.extend_from_slice(&0i32.to_be_bytes()); // scale = 0
-        data.push(0x01); // MSB of 256
-        data.push(0x00); // LSB of 256
-        let (remaining, value) = parse_decimal(&data).unwrap();
-        assert!(remaining.is_empty());
-        match value {
-            Value::Decimal {
-                scale,
-                ref unscaled,
-            } => {
-                assert_eq!(scale, 0);
-                assert_eq!(unscaled, &[0x01, 0x00], "unscaled [0x01,0x00] = 256");
-            }
-            other => panic!("Expected Decimal value, got {:?}", other),
-        }
-    }
-
-    #[test]
-    fn test_parse_inet_ipv4() {
-        // IPv4 address (4 bytes)
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(4)); // length = 4
-        data.extend_from_slice(&[192, 168, 1, 1]); // 192.168.1.1
-        let (remaining, value) = parse_inet(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Inet(vec![192, 168, 1, 1]));
-    }
-
-    #[test]
-    fn test_parse_inet_ipv6() {
-        // IPv6 address (16 bytes)
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(16)); // length = 16
-        let ipv6_bytes: [u8; 16] = [
-            0x20, 0x01, 0x0d, 0xb8, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-            0x00, 0x01,
-        ];
-        data.extend_from_slice(&ipv6_bytes); // 2001:db8::1
-        let (remaining, value) = parse_inet(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Inet(ipv6_bytes.to_vec()));
-    }
-
-    #[test]
-    fn test_parse_duration() {
-        // Duration: months (signed VInt) + days (signed VInt) + nanos (signed VInt)
-        // Returns Value::Duration { months, days, nanos } preserving calendar semantics.
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(1)); // 1 month
-        data.extend_from_slice(&encode_vint(15)); // 15 days
-        data.extend_from_slice(&encode_vint(3_600_000_000_000_i64)); // 1 hour in nanos
-        let (remaining, value) = parse_duration(&data).unwrap();
-        assert!(remaining.is_empty());
-        match value {
-            Value::Duration {
-                months,
-                days,
-                nanos,
-            } => {
-                assert_eq!(months, 1, "months should be 1");
-                assert_eq!(days, 15, "days should be 15");
-                assert_eq!(
-                    nanos, 3_600_000_000_000,
-                    "nanos should be 1 hour in nanoseconds"
-                );
-            }
-            other => panic!("Expected Duration value, got {:?}", other),
-        }
-    }
-
-    /// S2 regression: duration with all-zero components should produce Duration{{0,0,0}}.
-    #[test]
-    fn test_parse_duration_zero() {
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(0));
-        data.extend_from_slice(&encode_vint(0));
-        data.extend_from_slice(&encode_vint(0));
-        let (remaining, value) = parse_duration(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(
-            value,
-            Value::Duration {
-                months: 0,
-                days: 0,
-                nanos: 0
-            },
-            "zero duration should be Duration{{0,0,0}}"
-        );
-    }
-
-    /// S2 regression: duration with negative nanos (nanoseconds can be negative).
-    #[test]
-    fn test_parse_duration_negative_nanos() {
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(0));
-        data.extend_from_slice(&encode_vint(0));
-        data.extend_from_slice(&encode_vint(-1_i64));
-        let (remaining, value) = parse_duration(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(
-            value,
-            Value::Duration {
-                months: 0,
-                days: 0,
-                nanos: -1
-            },
-            "duration with -1 nanos should be Duration{{0,0,-1}}"
-        );
-    }
-
-    #[test]
-    fn test_parse_map_empty() {
-        // Empty map: count = 0
-        use super::super::vint::encode_vint;
-        let data = encode_vint(0); // count = 0
-        let (remaining, value) = parse_map(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Map(vec![]));
-    }
-
-    #[test]
-    fn test_parse_map_with_entries() {
-        // Map with 2 entries: int -> text
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(2)); // count = 2
-        data.push(CqlTypeId::Int as u8); // key type
-        data.push(CqlTypeId::Varchar as u8); // value type
-
-        // Entry 1: key=1, value="one"
-        data.extend_from_slice(&encode_vint(4)); // key length
-        data.extend_from_slice(&1i32.to_be_bytes()); // key = 1
-        data.extend_from_slice(&encode_vint(3)); // value length
-        data.extend_from_slice(b"one"); // value = "one"
-
-        // Entry 2: key=2, value="two"
-        data.extend_from_slice(&encode_vint(4)); // key length
-        data.extend_from_slice(&2i32.to_be_bytes()); // key = 2
-        data.extend_from_slice(&encode_vint(3)); // value length
-        data.extend_from_slice(b"two"); // value = "two"
-
-        let (remaining, value) = parse_map(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(
-            value,
-            Value::Map(vec![
-                (Value::Integer(1), Value::Text("one".to_string())),
-                (Value::Integer(2), Value::Text("two".to_string())),
-            ])
-        );
-    }
-
-    #[test]
-    fn test_parse_set_empty() {
-        // Empty set: count = 0
-        use super::super::vint::encode_vint;
-        let data = encode_vint(0); // count = 0
-        let (remaining, value) = parse_set(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Set(vec![]));
-    }
-
-    #[test]
-    fn test_parse_set_with_entries() {
-        // Set with 3 integer elements
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(3)); // count = 3
-        data.push(CqlTypeId::Int as u8); // element type
-
-        // Element 1: 10
-        data.extend_from_slice(&encode_vint(4)); // length
-        data.extend_from_slice(&10i32.to_be_bytes());
-
-        // Element 2: 20
-        data.extend_from_slice(&encode_vint(4)); // length
-        data.extend_from_slice(&20i32.to_be_bytes());
-
-        // Element 3: 30
-        data.extend_from_slice(&encode_vint(4)); // length
-        data.extend_from_slice(&30i32.to_be_bytes());
-
-        let (remaining, value) = parse_set(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(
-            value,
-            Value::Set(vec![
-                Value::Integer(10),
-                Value::Integer(20),
-                Value::Integer(30)
-            ])
-        );
-    }
-
-    #[test]
-    fn test_parse_date() {
-        // Cassandra DATE format: stored as u32 with Integer.MIN_VALUE offset
-        // For example, 2023-12-18 is 19710 days since epoch
-        // Cassandra stores it as: 19710 - i32::MIN = 19710 + 2147483648 = 2147503358
-        let days_since_epoch: i32 = 19710; // 2023-12-18
-        let stored = (days_since_epoch as u32).wrapping_sub(i32::MIN as u32);
-        let data = stored.to_be_bytes();
-        let (remaining, value) = parse_date(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Date(days_since_epoch));
-    }
-
-    #[test]
-    fn test_parse_time() {
-        // Time: nanoseconds since midnight (i64)
-        let nanos: i64 = 43_200_000_000_000; // 12:00:00 in nanoseconds
-        let data = nanos.to_be_bytes();
-        let (remaining, value) = parse_time(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Time(nanos));
-    }
-
-    #[test]
-    fn test_parse_cql_type_id_all_types() {
-        // Test all known type IDs
-        let type_ids = vec![
-            (0x00, CqlTypeId::Custom),
-            (0x01, CqlTypeId::Ascii),
-            (0x02, CqlTypeId::BigInt),
-            (0x03, CqlTypeId::Blob),
-            (0x04, CqlTypeId::Boolean),
-            (0x05, CqlTypeId::Counter),
-            (0x06, CqlTypeId::Decimal),
-            (0x07, CqlTypeId::Double),
-            (0x08, CqlTypeId::Float),
-            (0x09, CqlTypeId::Int),
-            (0x0B, CqlTypeId::Timestamp),
-            (0x0C, CqlTypeId::Uuid),
-            (0x0D, CqlTypeId::Varchar),
-            (0x0E, CqlTypeId::Varint),
-            (0x0F, CqlTypeId::Timeuuid),
-            (0x10, CqlTypeId::Inet),
-            (0x11, CqlTypeId::Date),
-            (0x12, CqlTypeId::Time),
-            (0x13, CqlTypeId::Smallint),
-            (0x14, CqlTypeId::Tinyint),
-            (0x15, CqlTypeId::Duration),
-            (0x20, CqlTypeId::List),
-            (0x21, CqlTypeId::Map),
-            (0x22, CqlTypeId::Set),
-            (0x30, CqlTypeId::Udt),
-            (0x31, CqlTypeId::Tuple),
-            (0xFF, CqlTypeId::Tombstone),
-        ];
-        for (byte, expected) in type_ids {
-            let data = [byte];
-            let (_, type_id) = parse_cql_type_id(&data).unwrap();
-            assert_eq!(type_id, expected, "Failed for byte 0x{:02X}", byte);
-        }
-    }
-
-    #[test]
-    fn test_parse_cql_type_id_invalid() {
-        // Test invalid type ID
-        let data = [0x0A]; // Not a valid type ID
-        let result = parse_cql_type_id(&data);
-        assert!(result.is_err(), "Should fail for invalid type ID 0x0A");
-    }
-
-    #[test]
-    fn test_parse_tinyint() {
-        let data = [0x7F]; // 127
-        let (remaining, value) = parse_tinyint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Integer(127));
-
-        let data = [0x80]; // -128 (signed)
-        let (remaining, value) = parse_tinyint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Integer(-128));
-    }
-
-    #[test]
-    fn test_parse_smallint() {
-        let data = 0x7FFFi16.to_be_bytes(); // 32767
-        let (remaining, value) = parse_smallint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Integer(32767));
-
-        let data = (-32768i16).to_be_bytes(); // -32768
-        let (remaining, value) = parse_smallint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Integer(-32768));
-    }
-
-    #[test]
-    fn test_parse_bigint() {
-        let data = 0x7FFF_FFFF_FFFF_FFFFi64.to_be_bytes(); // i64::MAX
-        let (remaining, value) = parse_bigint(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::BigInt(i64::MAX));
-    }
-
-    #[test]
-    fn test_parse_counter() {
-        // Counter is stored as i64 big-endian
-        let counter_value: i64 = 42;
-        let data = counter_value.to_be_bytes();
-        let (remaining, value) = parse_counter(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Counter(42));
-    }
-
-    #[test]
-    fn test_parse_float() {
-        let float_value: f32 = 42.5; // Use a non-constant value
-        let data = float_value.to_be_bytes();
-        let (remaining, value) = parse_float(&data).unwrap();
-        assert!(remaining.is_empty());
-        if let Value::Float32(f) = value {
-            assert!((f - 42.5_f32).abs() < 0.0001);
-        } else {
-            panic!("Expected Float32, got {:?}", value);
-        }
-    }
-
-    #[test]
-    fn test_parse_double() {
-        let double_value: f64 = std::f64::consts::E;
-        let data = double_value.to_be_bytes();
-        let (remaining, value) = parse_double(&data).unwrap();
-        assert!(remaining.is_empty());
-        if let Value::Float(f) = value {
-            assert!((f - std::f64::consts::E).abs() < 0.0001);
-        } else {
-            panic!("Expected Float");
-        }
-    }
-
-    #[test]
-    fn test_parse_map_with_null_value() {
-        // Map with null value (length = -1)
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(1)); // count = 1
-        data.push(CqlTypeId::Int as u8); // key type
-        data.push(CqlTypeId::Varchar as u8); // value type
-
-        // Entry: key=1, value=null
-        data.extend_from_slice(&encode_vint(4)); // key length
-        data.extend_from_slice(&1i32.to_be_bytes()); // key = 1
-        data.extend_from_slice(&encode_vint(-1)); // value = null
-
-        let (remaining, value) = parse_map(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::Map(vec![(Value::Integer(1), Value::Null)]));
-    }
-
-    #[test]
-    fn test_parse_list_with_null_element() {
-        // List with null element (length = -1)
-        use super::super::vint::encode_vint;
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(2)); // count = 2
-        data.push(CqlTypeId::Int as u8); // element type
-
-        // Element 1: 42
-        data.extend_from_slice(&encode_vint(4)); // length
-        data.extend_from_slice(&42i32.to_be_bytes());
-
-        // Element 2: null
-        data.extend_from_slice(&encode_vint(-1)); // null marker
-
-        let (remaining, value) = parse_list(&data).unwrap();
-        assert!(remaining.is_empty());
-        assert_eq!(value, Value::List(vec![Value::Integer(42), Value::Null]));
-    }
-
-    // Issue #264: Collection max elements limit test
-    #[test]
-    fn test_collection_max_elements_limit() {
-        use super::super::vint::encode_vint;
-
-        // Build list data claiming 1,000,001 elements (exceeds MAX_COLLECTION_SIZE)
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(1_000_001i64)); // Count > 1,000,000 limit
-        data.push(CqlTypeId::Int as u8); // Element type
-
-        let result = parse_list(&data);
-        assert!(
-            result.is_err(),
-            "Should reject list with > 1,000,000 elements"
-        );
-
-        // nom error - check it's TooLarge
-        if let Err(nom::Err::Error(e)) = result {
-            assert_eq!(
-                e.code,
-                nom::error::ErrorKind::TooLarge,
-                "Error should be TooLarge, got: {:?}",
-                e.code
-            );
-        } else {
-            panic!("Expected nom::Err::Error, got something else");
-        }
-    }
-
-    // Issue #264: Map max elements limit test
-    #[test]
-    fn test_map_max_elements_limit() {
-        use super::super::vint::encode_vint;
-
-        // Build map data claiming 1,000,001 entries
-        let mut data = Vec::new();
-        data.extend_from_slice(&encode_vint(1_000_001i64)); // Count > 1,000,000 limit
-        data.push(CqlTypeId::Int as u8); // Key type
-        data.push(CqlTypeId::Varchar as u8); // Value type
-
-        let result = parse_map(&data);
-        assert!(
-            result.is_err(),
-            "Should reject map with > 1,000,000 entries"
-        );
-
-        // nom error - check it's TooLarge (consistent with list test)
-        if let Err(nom::Err::Error(e)) = result {
-            assert_eq!(
-                e.code,
-                nom::error::ErrorKind::TooLarge,
-                "Error should be TooLarge, got: {:?}",
-                e.code
-            );
-        } else {
-            panic!("Expected nom::Err::Error, got something else");
-        }
-    }
-
-    #[test]
-    fn test_parse_cql_value_for_type_frozen_int() {
-        // 4 bytes of i32 BE representing the value 42
-        let data: &[u8] = &[0x00, 0x00, 0x00, 0x2A];
-        let cql_type = CqlType::Frozen(Box::new(CqlType::Int));
-        let result = parse_cql_value_for_type(data, &cql_type)
-            .expect("parse_cql_value_for_type should succeed for Frozen<Int>");
-        assert_eq!(result, Value::Frozen(Box::new(Value::Integer(42))));
-    }
-}
